fix-a-warn
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@318 406a0c4d-033a-0410-8de8-e80135713968

+++ src/search.cpp
@@ -131,7 +131,7 @@ int main ( int argc, char ** argv )
 			}
 		}
 
-		assert ( iPos<sizeof(sQuery) );
+		assert ( iPos<(int)sizeof(sQuery) );
 		sQuery[iPos] = '\0';
 	}
